requirements.txt
coinbase-advanced-py>=1.8.2
pandas>=2.2.0
python-dotenv>=1.0.1

